testing ethans user story
As a user I want my pet's satisfaction level to reset to 100% when it levels up, so the player feels rewarded for staying off social media

i was the tester and tested the code i connected this script to a gameobject and ran it. everything worked and passed the acceptance test.